Refactor: rename and reorganize participation-related classes for clarity

+++ src/main/java/unq/dapp/grupoj/soccergenius/repository/PlayerRepository.java
@@ -3,5 +3,5 @@ package unq.dapp.grupoj.soccergenius.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import unq.dapp.grupoj.soccergenius.model.player.Player;
 
-public interface PlayerRepository extends JpaRepository<Player, Long> {
+public interface PlayerRepository extends JpaRepository<Player, Integer> {
 }
